Refactor data_utils
